use SpringData Aerospike beans directly instead of duplicating them
