+++ spring-boot-autoconfigure-data-aerospike/src/main/resources/META-INF/spring.factories
@@ -2,6 +2,7 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 org.springframework.boot.autoconfigure.aerospike.AerospikeAutoConfiguration,\
 org.springframework.boot.autoconfigure.data.aerospike.AerospikeDataAutoConfiguration,\
+org.springframework.boot.autoconfigure.aerospike.AerospikeClientConfiguration,\
 org.springframework.boot.autoconfigure.data.aerospike.AerospikeRepositoriesAutoConfiguration,\
 org.springframework.boot.autoconfigure.data.aerospike.AerospikeReactiveDataAutoConfiguration,\
 org.springframework.boot.autoconfigure.data.aerospike.AerospikeReactiveRepositoriesAutoConfiguration
