+++ spring-boot-autoconfigure-data-aerospike/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,5 +1,6 @@
 org.springframework.boot.autoconfigure.aerospike.AerospikeAutoConfiguration
 org.springframework.boot.autoconfigure.data.aerospike.AerospikeDataAutoConfiguration
+org.springframework.boot.autoconfigure.aerospike.AerospikeClientConfiguration
 org.springframework.boot.autoconfigure.data.aerospike.AerospikeRepositoriesAutoConfiguration
 org.springframework.boot.autoconfigure.data.aerospike.AerospikeReactiveDataAutoConfiguration
 org.springframework.boot.autoconfigure.data.aerospike.AerospikeReactiveRepositoriesAutoConfiguration
